+++ packages/apps/shopify-app-remix/src/server/authenticate/admin/billing/types.ts
@@ -134,9 +134,11 @@ export interface BillingContext<Config extends AppConfigArg> {
    *   billing: {
    *     [MONTHLY_PLAN]: {
    *       lineItems: [
-   *         amount: 5,
-   *         currencyCode: 'USD',
-   *         interval: BillingInterval.Every30Days,
+   *         {
+   *           amount: 5,
+   *           currencyCode: 'USD',
+   *           interval: BillingInterval.Every30Days,
+   *         }
    *       ],
    *     },
    *   }
@@ -179,16 +181,20 @@ export interface BillingContext<Config extends AppConfigArg> {
    *   billing: {
    *     [MONTHLY_PLAN]: {
    *       lineItems: [
-   *         amount: 5,
-   *         currencyCode: 'USD',
-   *         interval: BillingInterval.Every30Days,
+   *         {
+   *           amount: 5,
+   *           currencyCode: 'USD',
+   *           interval: BillingInterval.Every30Days,
+   *         }
    *       ],
    *     },
    *     [ANNUAL_PLAN]: {
    *       lineItems: [
-   *         amount: 50,
-   *         currencyCode: 'USD',
-   *         interval: BillingInterval.Annual,
+   *         {
+   *           amount: 50,
+   *           currencyCode: 'USD',
+   *           interval: BillingInterval.Annual,
+   *         }
    *       ],
    *     },
    *   }
@@ -237,16 +243,20 @@ export interface BillingContext<Config extends AppConfigArg> {
    *   billing: {
    *     [MONTHLY_PLAN]: {
    *       lineItems: [
-   *         amount: 5,
-   *         currencyCode: 'USD',
-   *         interval: BillingInterval.Every30Days,
+   *         {
+   *           amount: 5,
+   *           currencyCode: 'USD',
+   *           interval: BillingInterval.Every30Days,
+   *         }
    *       ],
    *     },
    *     [ANNUAL_PLAN]: {
    *       lineItems: [
-   *         amount: 50,
-   *         currencyCode: 'USD',
-   *         interval: BillingInterval.Annual,
+   *         {
+   *           amount: 50,
+   *           currencyCode: 'USD',
+   *           interval: BillingInterval.Annual,
+   *         }
    *       ],
    *     },
    *   }
@@ -316,16 +326,20 @@ export interface BillingContext<Config extends AppConfigArg> {
    *   billing: {
    *     [MONTHLY_PLAN]: {
    *       lineItems: [
-   *         amount: 5,
-   *         currencyCode: 'USD',
-   *         interval: BillingInterval.Every30Days,
+   *         {
+   *           amount: 5,
+   *           currencyCode: 'USD',
+   *           interval: BillingInterval.Every30Days,
+   *         }
    *       ],
    *     },
    *     [ANNUAL_PLAN]: {
    *       lineItems: [
-   *         amount: 50,
-   *         currencyCode: 'USD',
-   *         interval: BillingInterval.Annual,
+   *         {
+   *           amount: 50,
+   *           currencyCode: 'USD',
+   *           interval: BillingInterval.Annual,
+   *         }
    *       ],
    *     },
    *   }
@@ -374,16 +388,20 @@ export interface BillingContext<Config extends AppConfigArg> {
    *   billing: {
    *     [MONTHLY_PLAN]: {
    *       lineItems: [
-   *         amount: 5,
-   *         currencyCode: 'USD',
-   *         interval: BillingInterval.Every30Days,
+   *         {
+   *           amount: 5,
+   *           currencyCode: 'USD',
+   *           interval: BillingInterval.Every30Days,
+   *         }
    *       ],
    *     },
    *     [ANNUAL_PLAN]: {
    *       lineItems: [
-   *         amount: 50,
-   *         currencyCode: 'USD',
-   *         interval: BillingInterval.Annual,
+   *         {
+   *           amount: 50,
+   *           currencyCode: 'USD',
+   *           interval: BillingInterval.Annual,
+   *         }
    *       ],
    *     },
    *   }
@@ -436,16 +454,20 @@ export interface BillingContext<Config extends AppConfigArg> {
    *   billing: {
    *     [MONTHLY_PLAN]: {
    *       lineItems: [
-   *         amount: 5,
-   *         currencyCode: 'USD',
-   *         interval: BillingInterval.Every30Days,
+   *         {
+   *           amount: 5,
+   *           currencyCode: 'USD',
+   *           interval: BillingInterval.Every30Days,
+   *         }
    *       ],
    *     },
    *     [ANNUAL_PLAN]: {
    *       lineItems: [
-   *         amount: 50,
-   *         currencyCode: 'USD',
-   *         interval: BillingInterval.Annual,
+   *         {
+   *           amount: 50,
+   *           currencyCode: 'USD',
+   *           interval: BillingInterval.Annual,
+   *         }
    *       ],
    *     },
    *   }
@@ -496,9 +518,11 @@ export interface BillingContext<Config extends AppConfigArg> {
    *   billing: {
    *     [USAGE_PLAN]: {
    *       lineItems: [
-   *         amount: 5,
-   *         currencyCode: 'USD',
-   *         interval: BillingInterval.Usage,
+   *         {
+   *           amount: 5,
+   *           currencyCode: 'USD',
+   *           interval: BillingInterval.Usage,
+   *         }
    *       ],
    *     },
    *   }