+++ packages/apps/session-storage/shopify-app-session-storage-drizzle/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @shopify/shopify-app-session-storage-drizzle
 
+## 2.0.7
+
+### Patch Changes
+
+- d2e3201: Updated `mysql2` dependencies
+
 ## 2.0.6
 
 ## 2.0.5