+++ packages/apps/session-storage/shopify-app-session-storage-kv/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @shopify/shopify-app-session-storage-kv
 
+## 4.0.7
+
 ## 4.0.6
 
 ## 4.0.5